Modify epoch time to double

+++ src/Licensing/LicenseChecker.cs
@@ -68,7 +68,7 @@ namespace Licensing {
                 RawData = licenseKey.Data,
                 Data = new LicenseKeyDataModel {
                     Email = splittedTextData[0],
-                    EpochTime = float.Parse(splittedTextData[1]),
+                    EpochTime = double.Parse(splittedTextData[1]),
                     Expiry = Int32.Parse(splittedTextData[2]),
                     Subscription = (SubscriptionTier) Int32.Parse(splittedTextData[3])
                 },
